code: detect base64url_decode error in oidc_proto_idtoken_validate_hash_value
oidc_util_base64url_decode can return -1, but the caller stored the
return value in an unsigned int, which turned the "<= 0" guard into a
test for == 0 only and silently let a NULL-input error slip through.
Use a signed int for decoded_len (matching the function's return type)
and cast when comparing against the unsigned hash_len.

+++ src/proto/id_token.c
@@ -252,16 +252,16 @@ static apr_byte_t oidc_proto_validate_hash(request_rec *r, const char *alg, cons
 
 	/* calculate the base64url-encoded value of the hash */
 	char *decoded = NULL;
-	unsigned int decoded_len = oidc_util_base64url_decode(r->pool, &decoded, hash);
+	int decoded_len = oidc_util_base64url_decode(r->pool, &decoded, hash);
 	if (decoded_len <= 0) {
 		oidc_error(r, "oidc_base64url_decode returned an error");
 		return FALSE;
 	}
 
-	oidc_debug(r, "hash_len=%d, decoded_len=%d, calc_len=%d", hash_len, decoded_len, calc_len);
+	oidc_debug(r, "hash_len=%u, decoded_len=%d, calc_len=%u", hash_len, decoded_len, calc_len);
 
 	/* compare the calculated hash against the provided hash */
-	if ((decoded_len != hash_len) || (calc_len < hash_len) || (memcmp(decoded, calc, hash_len) != 0)) {
+	if (((unsigned int)decoded_len != hash_len) || (calc_len < hash_len) || (memcmp(decoded, calc, hash_len) != 0)) {
 		oidc_error(r, "provided \"%s\" hash value (%s) does not match the calculated value", type, hash);
 		return FALSE;
 	}
